Stop building an abandoned project.
Upstream has moved on and requires additional dependencies, and there's
no obvious sign that anyone's touched this since 2013 other than to try
to keep it building. Let's see if we can get out of that business.

Anyone wanting to bring this back to life should be prepared to own
importing the other dependencies, and getting this back in sync with
upstream.

Test: treehugger
Change-Id: I85fc519d928cd783617583568e799f24e6dbe226